backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for PAIM project APP"
authors = ["Wiktor Frątczak <01171649@pw.edu.pl>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "application"},
    {include = "domain"},
    {include = "infrastructure"},
    {include = "*.env"}]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"
pre-commit = "^4.0.1"
mongo = "^0.2.0"
python-multipart = "^0.0.19"
pytest = "^8.3.4"
httpx = "^0.28.1"
dependency-injector = "^4.44.0"
argon2-cffi = "^23.1.0"
python-jose = "^3.3.0"
bcrypt = "^4.2.1"
passlib = "^1.7.4"
asyncio = "^3.4.3"
pydantic = {extras = ["email"], version = "^2.10.5"}
stripe = "^11.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
black = "^24.10.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
